Major Color Changes (#111)
* Major Color Changes

1. Changed colors for tags for Project and Resource sections.
2. Changed the font color and backgrounds to be navy blue.
3.

* Student Organization Colored Tags

1. Added color to the tags of the Student Organizations.
2. Added 2 additional colors to all 3 sections of Student Organizations, Projects and Resources.

* Changed from upper to lower case

1. Changed to lowercase to satisfy lint test.

* Made font in tags bold.

In the student organization, project, and resource page, added a css style to make it bold to make it easier to see.

* Fixed Lint Test Error

Placed font-weight below display to fix the linting error. Removed styling in organization scss. There is a minor bug with the cards being called resource card there to be fixed later.

* Fixed linting error issue

* Removed light pink & lilac

Removed 2 similar pinkish colors.